refactor: move types to src/types, add @customTypes alias
