+++ src/App/pages/Product/Product.tsx
@@ -5,7 +5,8 @@ import Loader, { LoaderSize } from "@components/Loader";
 import Wrapper from "@components/Wrapper";
 import { DEFAULT_ERROR_STATUS, DEFAULT_PRODUCT_ID } from "@config/constants";
 import { productsMock } from "@config/data";
-import { ApiError, Product as ProductType } from "@config/types";
+import ApiError from "@customTypes/ApiError";
+import ProductType from "@customTypes/Product";
 import useFetchProducts, { FetchFunctionParams } from "@hooks/useFetchProducts";
 import { getProductById } from "@services/products";
 import { useParams } from "react-router-dom";
