+++ src/types/ApiError.ts
@@ -1,0 +1,6 @@
+type ApiError = {
+  code: number;
+  message: string;
+};
+
+export default ApiError;
