+++ src/types/Category.ts
@@ -1,0 +1,7 @@
+type Category = {
+  id: number;
+  name: string;
+  image: string;
+};
+
+export default Category;
