+++ src/types/Endpoint.ts
@@ -1,0 +1,6 @@
+type Endpoint = {
+  base: string;
+  [key: string]: string;
+};
+
+export default Endpoint;
