+++ src/types/HeaderNavItem.ts
@@ -1,0 +1,8 @@
+import { Routes } from "@config/constants";
+
+type HeaderNavItem = {
+  text: string;
+  link: Routes;
+};
+
+export default HeaderNavItem;
