+++ src/types/HeroContent.ts
@@ -1,0 +1,3 @@
+type HeroContent = { heading: string; caption: string };
+
+export default HeroContent;
