+++ src/types/Pages.ts
@@ -1,0 +1,7 @@
+enum Pages {
+  MAIN = "main",
+  CATEGORIES = "categories",
+  ABOUT = "about",
+}
+
+export default Pages;
