+++ src/types/Product.ts
@@ -1,0 +1,12 @@
+import Category from "@customTypes/Category";
+
+type Product = {
+  id: number;
+  title: string;
+  price: number;
+  description: string;
+  category: Category;
+  images: string[];
+};
+
+export default Product;
