+++ src/utils/renderProductCards.tsx
@@ -1,6 +1,6 @@
 import Card from "@components/Card";
 import Price from "@components/Price";
-import { Product } from "@config/types";
+import Product from "@customTypes/Product";
 
 const renderProductCards = (products: Product[]) => {
   return products.map((product) => {
